+++ clients/python/tests/test_marquez_client.py
@@ -11,7 +11,9 @@ from datetime import timedelta, date
 from marquez_client import MarquezClient
 from marquez_client.constants import (
     DEFAULT_LIMIT,
-    DEFAULT_OFFSET
+    DEFAULT_OFFSET,
+    DEFAULT_DEPTH,
+    DEFAULT_WITH_DOWNSTREAM
 )
 from marquez_client.models import (
     DatasetType,
@@ -236,6 +238,37 @@ EVENT = {
     'outputs': [],
     'producer': PRODUCER
 }
+
+COLUMN_LINEAGE = [
+    {
+        'id': 'datasetField:namespace:commonDataset:columnA',
+        'type': 'DATASET_FIELD',
+        'data': {
+            'type': 'DATASET_FIELD',
+            'namespace': 'namespace',
+            'dataset': 'otherDataset',
+            'field': 'columnA',
+            'fieldType': 'integer',
+            'transformationDescription': 'identical',
+            'transformationType': 'IDENTITY',
+            'inputFields': [
+                {'namespace': 'namespace', 'dataset':  'otherDataset', 'field': 'columnB'}
+            ]
+        },
+        'inEdges': [
+            {
+                'origin': 'datasetField:namespace:otherDataset:columnB',
+                'destination': 'datasetField:namespace:commonDataset:columnA'
+            }
+        ],
+        'outEdges': [
+            {
+                'origin': 'datasetField:namespace:commonDataset:columnA',
+                'destination': 'datasetField:namespace:otherDataset:columnC'
+            }
+        ]
+    }
+]
 
 
 @pytest.fixture
@@ -935,6 +968,91 @@ def test_list_tags(mock_get, client):
         params={
             'limit': DEFAULT_LIMIT,
             'offset': DEFAULT_OFFSET
+        },
+        timeout=mock.ANY
+    )
+
+
+@mock.patch('requests.get')
+def test_get_column_lineage_by_dataset(mock_get, client):
+    mock_get.return_value.status_code.return_value = HTTPStatus.OK
+    mock_get.return_value.json.return_value = COLUMN_LINEAGE
+
+    column_lineage = client.get_column_lineage_by_dataset(
+        "namespace_a",
+        "dataset_a",
+        DEFAULT_DEPTH,
+        DEFAULT_WITH_DOWNSTREAM
+    )
+
+    assert column_lineage == COLUMN_LINEAGE
+
+    mock_get.assert_called_once_with(
+        url=client._url(
+            '/column-lineage'
+        ),
+        headers=mock.ANY,
+        params={
+            'nodeId': 'dataset:namespace_a:dataset_a',
+            'depth': DEFAULT_DEPTH,
+            'withDownstream': DEFAULT_WITH_DOWNSTREAM
+        },
+        timeout=mock.ANY
+    )
+
+
+@mock.patch('requests.get')
+def test_get_column_lineage_by_dataset_field(mock_get, client):
+    mock_get.return_value.status_code.return_value = HTTPStatus.OK
+    mock_get.return_value.json.return_value = COLUMN_LINEAGE
+
+    column_lineage = client.get_column_lineage_by_dataset_field(
+        "namespace_a",
+        "dataset_a",
+        "field_a",
+        DEFAULT_DEPTH,
+        DEFAULT_WITH_DOWNSTREAM
+    )
+
+    assert column_lineage == COLUMN_LINEAGE
+
+    mock_get.assert_called_once_with(
+        url=client._url(
+            '/column-lineage'
+        ),
+        headers=mock.ANY,
+        params={
+            'nodeId': 'datasetField:namespace_a:dataset_a:field_a',
+            'depth': DEFAULT_DEPTH,
+            'withDownstream': DEFAULT_WITH_DOWNSTREAM
+        },
+        timeout=mock.ANY
+    )
+
+
+@mock.patch('requests.get')
+def test_get_column_lineage_by_job(mock_get, client):
+    mock_get.return_value.status_code.return_value = HTTPStatus.OK
+    mock_get.return_value.json.return_value = COLUMN_LINEAGE
+
+    column_lineage = client.get_column_lineage_by_job(
+        "namespace_a",
+        "job_a",
+        DEFAULT_DEPTH,
+        DEFAULT_WITH_DOWNSTREAM
+    )
+
+    assert column_lineage == COLUMN_LINEAGE
+
+    mock_get.assert_called_once_with(
+        url=client._url(
+            '/column-lineage'
+        ),
+        headers=mock.ANY,
+        params={
+            'nodeId': 'job:namespace_a:job_a',
+            'depth': DEFAULT_DEPTH,
+            'withDownstream': DEFAULT_WITH_DOWNSTREAM
         },
         timeout=mock.ANY
     )